Add symbol for `CdMix`
* `CdMix` calls into a function that does not currently have an
entry in the symbol table. This was producing undefined reference
errors during linking, so the `undefined_funcs_auto.txt` file was
added as a linkerscript.

* Organized segments back into a single `code` segment. It was
nice to have correct coverage statistics, but there were issues
with segment alignment at link time which made the tradeoff not
worth it. It's probably easier to patch `splat` to fix the
reporting.